Add a simple Flask api and template to push/pull data
This adds a `GET` and `POST` API on `/`, along with a template and form which
allows a user to select options for the features in our logistic model.

There is currently tight coupling between the HTML and what the model sample data
is using in the interest of time, but we may want to have a mapping for these
values eventually.

The form/API handles errors for null values. There is a mock currently which
will eventually be the call to our module to run our data through the prediction
model.

I've updated the README to reflect the steps needed to install the dependencies
of the Flask app and start the server.