Add mortality predictor
This adds a `predictor` module which we'll use to push data from the API into our
predictive model.

The module loads the mortality model and runs our form data through it. Improvements
could be made to ensure we've validating/simplifying data checks, but this is a
first pass.

We'll be able to add the new hospitalization model into this module and output
to the UI as well.